arena: fix afree() bug
afree() was unable to find the BMeta header before the data passed to
it, causing find_free_block() to be unable to do so, even when a free
block should have been present. this was because the pointer passed to
afree() was not reinterpreted as bytes, causing the pointer arithmetic
to be incorrect.