parser: start using arrays instead of linked lists...
instead of the numerous linked lists the parser previously used, it
now uses arrays to deal with expression bodies and lists. this allows
for much easier traversal of nested expression, and has paved the way
for a custom array type in the future to manage length and
such. furthermore, the analyzer is now able to mostly construct a
working symbol table.